backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.9.0